requirements.txt
# Django and REST Framework
Django==4.2.26
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.12.0
drf-spectacular==0.27.1
django-filter==24.2
whitenoise==6.6.0

# WSGI Server for production
gunicorn==22.0.0

# Database
psycopg2-binary==2.9.9

# Development tools
pytest==8.3.3
pytest-django==4.9.0
flake8==7.0.0
black==24.3.0
